+++ src/Http/SaloonRequest.php
@@ -3,17 +3,17 @@
 namespace Sammyjo20\Saloon\Http;
 
 use ReflectionClass;
-use Sammyjo20\Saloon\Traits\CollectsData;
-use Sammyjo20\Saloon\Traits\SendsRequests;
-use Sammyjo20\Saloon\Traits\CollectsConfig;
-use Sammyjo20\Saloon\Traits\CollectsHeaders;
-use Sammyjo20\Saloon\Traits\CollectsHandlers;
-use Sammyjo20\Saloon\Traits\CollectsQueryParams;
-use Sammyjo20\Saloon\Traits\CollectsInterceptors;
-use Sammyjo20\Saloon\Traits\Features\HasResponseClass;
-use Sammyjo20\Saloon\Interfaces\SaloonRequestInterface;
-use Sammyjo20\Saloon\Exceptions\SaloonMethodNotFoundException;
 use Sammyjo20\Saloon\Exceptions\SaloonInvalidConnectorException;
+use Sammyjo20\Saloon\Exceptions\SaloonMethodNotFoundException;
+use Sammyjo20\Saloon\Interfaces\SaloonRequestInterface;
+use Sammyjo20\Saloon\Traits\CollectsConfig;
+use Sammyjo20\Saloon\Traits\CollectsData;
+use Sammyjo20\Saloon\Traits\CollectsHandlers;
+use Sammyjo20\Saloon\Traits\CollectsHeaders;
+use Sammyjo20\Saloon\Traits\CollectsInterceptors;
+use Sammyjo20\Saloon\Traits\CollectsQueryParams;
+use Sammyjo20\Saloon\Traits\HasCustomResponses;
+use Sammyjo20\Saloon\Traits\SendsRequests;
 
 abstract class SaloonRequest implements SaloonRequestInterface
 {
@@ -23,7 +23,7 @@ abstract class SaloonRequest implements SaloonRequestInterface
         CollectsConfig,
         CollectsHandlers,
         CollectsInterceptors,
-        HasResponseClass,
+        HasCustomResponses,
         SendsRequests;
 
     /**
